# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ecards/{id}/extend": {
      "post": {
        "operationId": "EcardsController_extend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendEcardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Prolonger l’échéance d’une de MES e-cards ACTIVE (D-026 : le créateur en a le droit — prolonger ne crée aucune valeur, cela retarde son propre remboursement).",
        "tags": [
          "ecards"
        ]
      }
    },
    "/admin/ecards": {
      "get": {
        "description": "La réponse ne contient AUCUN code, y compris quand la recherche s’est faite par code : un code est de la valeur au porteur. Chercher n’est pas restituer.",
        "operationId": "EcardsAdminController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "code",
            "required": false,
            "in": "query",
            "description": "Code d’e-card, en correspondance EXACTE. La réponse ne contient JAMAIS de code (voir EcardAdminRowDto).",
            "schema": {
              "maxLength": 32,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "USED",
                "REVOKED",
                "EXPIRED"
              ]
            }
          },
          {
            "name": "origin",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "MEMBER",
                "GENESIS"
              ]
            }
          },
          {
            "name": "creatorMemberId",
            "required": false,
            "in": "query",
            "description": "Cartes créées par ce membre.",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "userMemberId",
            "required": false,
            "in": "query",
            "description": "Cartes consommées par ce membre.",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Créées à partir de cette date (incluse), au format ISO.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Créées jusqu’à cette date (incluse), au format ISO.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "valueDt",
                "expiresAt",
                "usedAt"
              ]
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EcardAdminPageDto"
                }
              }
            }
          }
        },
        "summary": "Lister les e-cards (filtres statut / origine / créateur / bénéficiaire / période, recherche par code EXACT, tri, pagination).",
        "tags": [
          "ecards-admin"
        ]
      }
    },
    "/admin/ecards/{id}": {
      "get": {
        "operationId": "EcardsAdminController_one",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EcardAdminDetailDto"
                }
              }
            }
          }
        },
        "summary": "Fiche d’une e-card : traçabilité création → utilisation, ce qu’elle a payé (commande ou adhésion), mouvements de solde liés.",
        "tags": [
          "ecards-admin"
        ]
      }
    },
    "/admin/ecards/genesis": {
      "post": {
        "description": "SEULE réponse de l’API admin qui contient un code en clair, et une seule fois : une carte de genèse n’a pas de créateur à qui demander son code, et aucun autre canal ne permettrait de le transmettre. Il n’est ensuite plus jamais consultable.",
        "operationId": "EcardsAdminController_genesis",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenesisEcardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenesisEcardResponseDto"
                }
              }
            }
          }
        },
        "summary": "Générer une e-card ex nihilo (amorçage / promo). Aucun solde débité ; expiration/révocation ne rembourse personne.",
        "tags": [
          "ecards-admin"
        ]
      }
    },
    "/admin/ecards/{id}/revoke": {
      "post": {
        "operationId": "EcardsAdminController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeEcardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EcardAdminActionDto"
                }
              }
            }
          }
        },
        "summary": "Révoquer une e-card ACTIVE : statut REVOKED, valeur (DT) RECRÉDITÉE au créateur — sauf carte de genèse, qui n’a débité personne.",
        "tags": [
          "ecards-admin"
        ]
      }
    },
    "/admin/ecards/{id}/extend": {
      "post": {
        "operationId": "EcardsAdminController_extend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendEcardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EcardAdminActionDto"
                }
              }
            }
          }
        },
        "summary": "Prolonger l’échéance de n’importe quelle e-card ACTIVE (support, D-026). Borné à 365 jours.",
        "tags": [
          "ecards-admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Najah Partners API",
    "description": "API backend — plateforme MLM (BV, e-cards, commissions)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ExtendEcardDto": {
        "type": "object",
        "properties": {
          "days": {
            "type": "number",
            "minimum": 1,
            "maximum": 365,
            "description": "Nombre de jours dont l’échéance est repoussée (depuis l’échéance courante, ou depuis maintenant si elle est déjà passée).",
            "example": 30
          }
        },
        "required": [
          "days"
        ]
      },
      "EcardAdminRowDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Identifiant technique — le seul désignateur affichable."
          },
          "valueDt": {
            "type": "string",
            "example": "2100.000",
            "description": "DINARS — une e-card est de l’argent (D-028)."
          },
          "status": {
            "enum": [
              "ACTIVE",
              "USED",
              "REVOKED",
              "EXPIRED"
            ],
            "type": "string"
          },
          "origin": {
            "enum": [
              "MEMBER",
              "GENESIS"
            ],
            "type": "string",
            "description": "MEMBER : la valeur a été débitée du solde du créateur. GENESIS : créée ex nihilo par un SUPER_ADMIN — personne n’est remboursé à son expiration."
          },
          "creatorMemberId": {
            "type": "number",
            "nullable": true,
            "description": "Créateur — `null` si et seulement si l’origine est GENESIS."
          },
          "creatorMemberCode": {
            "type": "string",
            "nullable": true
          },
          "creatorName": {
            "type": "string",
            "nullable": true
          },
          "userMemberId": {
            "type": "number",
            "nullable": true,
            "description": "Bénéficiaire qui l’a consommée — `null` tant qu’elle n’est pas USED."
          },
          "userMemberCode": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "usedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "`null` = illimité (paramètre à -1)."
          },
          "closedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "Passage à EXPIRED / REVOKED."
          },
          "orderId": {
            "type": "number",
            "nullable": true,
            "description": "Commande réglée par cette carte (D-041) — exclusif avec l’adhésion."
          },
          "orderContext": {
            "nullable": true,
            "enum": [
              "ACTIVATION",
              "FREE"
            ],
            "type": "string"
          },
          "membershipPaymentId": {
            "type": "number",
            "nullable": true,
            "description": "Paiement d’adhésion réglé par cette carte — exclusif avec la commande."
          },
          "membershipPaymentType": {
            "nullable": true,
            "enum": [
              "REGISTRATION",
              "RENEWAL"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "valueDt",
          "status",
          "origin",
          "createdAt"
        ]
      },
      "EcardAdminPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EcardAdminRowDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "totalValueDt": {
            "type": "string",
            "example": "12500.000",
            "description": "DINARS — somme des valeurs des cartes filtrées (pas seulement de la page)."
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize",
          "totalValueDt"
        ]
      },
      "EcardLedgerRowDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "memberId": {
            "type": "number"
          },
          "type": {
            "type": "string"
          },
          "amountDt": {
            "type": "string",
            "example": "-2100.000",
            "description": "DINARS, signé."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "memberId",
          "type",
          "amountDt",
          "createdAt"
        ]
      },
      "EcardAdminDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Identifiant technique — le seul désignateur affichable."
          },
          "valueDt": {
            "type": "string",
            "example": "2100.000",
            "description": "DINARS — une e-card est de l’argent (D-028)."
          },
          "status": {
            "enum": [
              "ACTIVE",
              "USED",
              "REVOKED",
              "EXPIRED"
            ],
            "type": "string"
          },
          "origin": {
            "enum": [
              "MEMBER",
              "GENESIS"
            ],
            "type": "string",
            "description": "MEMBER : la valeur a été débitée du solde du créateur. GENESIS : créée ex nihilo par un SUPER_ADMIN — personne n’est remboursé à son expiration."
          },
          "creatorMemberId": {
            "type": "number",
            "nullable": true,
            "description": "Créateur — `null` si et seulement si l’origine est GENESIS."
          },
          "creatorMemberCode": {
            "type": "string",
            "nullable": true
          },
          "creatorName": {
            "type": "string",
            "nullable": true
          },
          "userMemberId": {
            "type": "number",
            "nullable": true,
            "description": "Bénéficiaire qui l’a consommée — `null` tant qu’elle n’est pas USED."
          },
          "userMemberCode": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "usedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "`null` = illimité (paramètre à -1)."
          },
          "closedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "Passage à EXPIRED / REVOKED."
          },
          "orderId": {
            "type": "number",
            "nullable": true,
            "description": "Commande réglée par cette carte (D-041) — exclusif avec l’adhésion."
          },
          "orderContext": {
            "nullable": true,
            "enum": [
              "ACTIVATION",
              "FREE"
            ],
            "type": "string"
          },
          "membershipPaymentId": {
            "type": "number",
            "nullable": true,
            "description": "Paiement d’adhésion réglé par cette carte — exclusif avec la commande."
          },
          "membershipPaymentType": {
            "nullable": true,
            "enum": [
              "REGISTRATION",
              "RENEWAL"
            ],
            "type": "string"
          },
          "ledgerEntries": {
            "description": "Mouvements de solde liés. RAPPEL D-025 : la CONSOMMATION n’écrit rien ici — aucun solde ne bouge quand une carte paie. Seules la création et le remboursement apparaissent.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EcardLedgerRowDto"
            }
          }
        },
        "required": [
          "id",
          "valueDt",
          "status",
          "origin",
          "createdAt",
          "ledgerEntries"
        ]
      },
      "GenesisEcardDto": {
        "type": "object",
        "properties": {
          "valueDt": {
            "type": "number",
            "minimum": 0.001,
            "description": "Valeur en DINARS de l’e-card générée ex nihilo (amorçage du réseau / promotion). Aucun solde n’est débité. 3 décimales au maximum (le millime).",
            "example": 2200
          },
          "expirationDays": {
            "type": "number",
            "minimum": -1,
            "description": "Durée de validité en jours ; -1 = illimité. Omis : le paramètre système ecard_expiration_days s’applique.",
            "example": 90
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "description": "Motif OBLIGATOIRE (promo, amorçage…) — tracé dans l’AuditLog.",
            "example": "Promotion lancement — e-card Silver"
          }
        },
        "required": [
          "valueDt",
          "reason"
        ]
      },
      "GenesisEcardResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "code": {
            "type": "string",
            "example": "HHD-7Z7-JJD-77D",
            "description": "Le code EN CLAIR, rendu une seule fois. Ne jamais le journaliser, ni le stocker côté client."
          },
          "valueDt": {
            "type": "string",
            "example": "2100.000"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "USED",
              "REVOKED",
              "EXPIRED"
            ],
            "type": "string"
          },
          "origin": {
            "enum": [
              "MEMBER",
              "GENESIS"
            ],
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "code",
          "valueDt",
          "status",
          "origin",
          "createdAt"
        ]
      },
      "RevokeEcardDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "description": "Motif OBLIGATOIRE (tracé dans l’AuditLog).",
            "example": "Carte transmise par erreur au mauvais bénéficiaire"
          }
        },
        "required": [
          "reason"
        ]
      },
      "EcardAdminActionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "valueDt": {
            "type": "string",
            "example": "2100.000"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "USED",
              "REVOKED",
              "EXPIRED"
            ],
            "type": "string"
          },
          "origin": {
            "enum": [
              "MEMBER",
              "GENESIS"
            ],
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "usedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "closedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "valueDt",
          "status",
          "origin",
          "createdAt"
        ]
      },
...
    }
  }
}